# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: DraftFly API — AI-powered reply automation for B2B sales agencies
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all client accounts
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a client account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a single client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ---------- PERSONAS ----------
...
components:
  schemas:
...
    Client:
      type: object
      required: [id, name, mode, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        company:
          type: ["string", "null"]
        # Optional: a client approving in the dashboard needs no Slack channel.
        slackChannel:
          type: ["string", "null"]
        slackWorkspaceId:
          type: ["string", "null"]
        slackBotToken:
          type: ["string", "null"]
        mode:
          type: string
          enum: [draft, auto]
        # Agency is absent on purpose: it means several Lemlist accounts under
        # one customer, which this model cannot express.
        plan:
          type: string
          enum: [starter, growth]
        lemlistApiKey:
          type: ["string", "null"]
        n8nWebhookUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        # Computed, display-ready billing state (see api-server lib/billing
        # accessStateFor). "managed" = operator-billed, never expires.
        billingStatus:
          type: string
          enum: [managed, trial, active, past_due, locked]
        # Days remaining when billingStatus is "trial", else null.
        trialDaysLeft:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    ClientInput:
      type: object
      required: [name, mode]
      properties:
        name:
          type: string
        company:
          type: string
        slackChannel:
          type: string
          pattern: "^[CG][A-Z0-9]{9,}$"
          description: >-
            Optional Slack channel ID starting with C or G, e.g. C012AB3CD45.
            Omit it for a client who approves drafts in the dashboard; they
            simply get no Slack approval card.
        slackWorkspaceId:
          type: string
        slackBotToken:
          type: string
        mode:
          type: string
          enum: [draft, auto]
        plan:
          type: string
          enum: [starter, growth]
        lemlistApiKey:
          type: string
        n8nWebhookUrl:
          type: string
    ClientUpdate:
      type: object
      properties:
        name:
          type: string
        company:
          type: string
        # Nullable so a channel can be removed. Without null there is no way to
        # express "this client no longer posts to Slack": omitting the field
        # leaves the old channel in place, and "" fails the pattern.
        slackChannel:
          type: ["string", "null"]
          pattern: "^[CG][A-Z0-9]{9,}$"
          description: "Slack channel ID starting with C or G, e.g. C012AB3CD45. Null removes it."
        slackWorkspaceId:
          type: string
        slackBotToken:
          type: string
        mode:
          type: string
          enum: [draft, auto]
        plan:
          type: string
          enum: [starter, growth]
        lemlistApiKey:
          type: string
        n8nWebhookUrl:
          type: string
        isActive:
          type: boolean

    # ---- Persona ----
...